requirements.txt
langchain
langchain_huggingface
langchain_openai
langgraph
torch
rich
unstructured[pdf] 
chromadb 
sentence-transformers 
pypdf 
langchain_experimental 
rank_bm25
modelscope
jieba

# Haruhi
transformers_stream_generator==0.0.4
einops
json_repair

# graph-based rag
setuptools==75.8.0
pyfiglet
loguru
anthropic
igraph
lazy-object-proxy
llama-index
llama-index-embeddings-ollama
llama-index-vector-stores-faiss
llama-index-vector-stores-postgres
llama-index-legacy
llama-index-embeddings-huggingface
faiss-gpu
numpy==1.26.4
umap
pcst_fast
huggingface_hub
graspologic
future